pyproject.toml
[project]
name = "vs-market"
version = "0.1.0"
description = ""
authors = [
    {name = "UK"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "bs4 (==0.0.2)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "python-dotenv (==1.0.1)",
    "requests (>=2.32.3,<3.0.0)",
    "lxml (==5.3.0)",
    "sqlalchemy (==2.0.37)",
    "uvicorn (==0.34.0)",
    "websockets (==14.2)",
    "redis (>=5.2.1,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "babel (>=2.17.0,<3.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "phonenumbers (>=8.13.54,<9.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "confluent-kafka (>=2.8.0,<3.0.0)",
    "edge-tts (>=7.0.0,<8.0.0)",
    "ulid-py (>=1.1.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
